\name{Ruby-package}
\alias{Ruby-package}
\alias{Ruby}
\docType{package}
\title{
\packageTitle{Ruby}
}
\description{
\packageDescription{Ruby}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Ruby}
\packageIndices{Ruby}
 See man pages,
}
\author{
\packageAuthor{Ruby}

Maintainer: \packageMaintainer{Ruby}
}
\references{
Evaluation and Comparison of Clustering Algorithms in Analyzing ES Cell Gene Expression Data

Are clusters found in one dataset present in another dataset

}


\keyword{ package }
\seealso{

}
\examples{
#
}
